// src/types/glicogest.types.ts
export type MeasurementType = 'JEJUM' | 'POS_CAFE' | 'POS_ALMOCO' | 'POS_JANTAR';